fix(backend): V1 마이그레이션 HC 트리거/함수 제거 및 레거시 엔티티 ID 타입 수정
Flyway V1 마이그레이션의 Heroku Connect 트리거(30개)와 함수(35개)를 제거하여
로컬 fresh DB에서 function does not exist 에러를 해결.
레거시 salesforce2 테이블의 serial4(INTEGER) 컬럼에 맞게 엔티티 ID를
Long → Int로 변경하여 Hibernate schema validation 에러를 해결.

+++ backend/src/main/kotlin/com/otoki/internal/common/service/MyStoreService.kt
@@ -111,7 +111,7 @@ class MyStoreService(
 
         return when {
             account != null -> MyStoreInfo(
-                storeId = account.id,
+                storeId = account.id.toLong(),
                 storeName = account.name ?: "",
                 storeCode = account.externalKey ?: "",
                 address = account.address1,
