+++ backend/src/main/kotlin/com/otoki/internal/entity/PushMessageReceiver.kt
@@ -15,7 +15,7 @@ class PushMessageReceiver(
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    val id: Long = 0,
+    val id: Int = 0,
 
     @Column(name = "sfid", length = 18)
     val sfid: String? = null,
