+++ backend/src/main/kotlin/com/otoki/internal/repository/AgreementRepository.kt
@@ -3,4 +3,4 @@ package com.otoki.internal.repository
 import com.otoki.internal.entity.Agreement
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface AgreementRepository : JpaRepository<Agreement, Long>
+interface AgreementRepository : JpaRepository<Agreement, Int>
