+++ backend/src/main/kotlin/com/otoki/internal/repository/HqReviewRepository.kt
@@ -3,4 +3,4 @@ package com.otoki.internal.repository
 import com.otoki.internal.entity.HqReview
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface HqReviewRepository : JpaRepository<HqReview, Long>
+interface HqReviewRepository : JpaRepository<HqReview, Int>
