+++ backend/src/main/kotlin/com/otoki/internal/repository/PushMessageReceiverRepository.kt
@@ -3,4 +3,4 @@ package com.otoki.internal.repository
 import com.otoki.internal.entity.PushMessageReceiver
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface PushMessageReceiverRepository : JpaRepository<PushMessageReceiver, Long>
+interface PushMessageReceiverRepository : JpaRepository<PushMessageReceiver, Int>
