+++ backend/src/main/kotlin/com/otoki/internal/repository/PushMessageRepository.kt
@@ -3,4 +3,4 @@ package com.otoki.internal.repository
 import com.otoki.internal.entity.PushMessage
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface PushMessageRepository : JpaRepository<PushMessage, Long>
+interface PushMessageRepository : JpaRepository<PushMessage, Int>
