+++ backend/src/test/kotlin/com/otoki/internal/common/service/HomeServiceTest.kt
@@ -415,7 +415,7 @@ class HomeServiceTest {
     }
 
     private fun createAccount(
-        id: Long = 0L,
+        id: Int = 0,
         sfid: String? = null,
         name: String? = null
     ): Account {
